requirements.txt
d3dshot~=0.1.5
numpy~=1.19.5
mouse~=0.7.1
Flask~=2.0.3
Pillow~=8.4.0
